Variaveis comp e troca trocadas de int para long

+++ bolhadireta.c
@@ -3,7 +3,7 @@
 #include <stdlib.h>
 #include <time.h>
 
-void bubbleSort(int* vet, int tam, int* troca, int* comp) {
+void bubbleSort(int* vet, int tam, long* troca, long* comp) {
     int i, j, aux;
     for (i = tam - 1; i >= 1; i--) {
         for (j = 0; j < i; j++) {
