+++ insbinaria.h
@@ -1,6 +1,6 @@
 #ifndef INSBINARIA_H
 #define INSBINARIA_H
 
-void insertionSortBin (int* vet, int tam, int* troca, int* comp);
+void insertionSortBin (int* vet, int tam, long* troca, long* comp);
 
 #endif
