+++ quicksortmediana.h
@@ -1,6 +1,6 @@
 #ifndef QUICKSORTMEDIANA_H
 #define QUICKSORTMEDIANA_H
 
-void quickSortMediana (int* vet, int esq, int dir, int* comp, int* troca);
+void quickSortMediana (int* vet, int esq, int dir, long* comp, long* troca);
 
 #endif
